+++ hahomematic/client/__init__.py
@@ -7,7 +7,8 @@ from datetime import datetime
 import logging
 from typing import Any, Final, cast
 
-from hahomematic import central_unit as hmcu
+from hahomematic import central as hmcu
+from hahomematic.client.xml_rpc import XmlRpcProxy
 from hahomematic.config import CALLBACK_WARN_INTERVAL, RECONNECT_WAIT
 from hahomematic.const import (
     EVENT_AVAILABLE,
@@ -24,22 +25,16 @@ from hahomematic.const import (
     HmParamsetKey,
     HmProductGroup,
     HmProxyInitState,
-)
-from hahomematic.decorators import measure_execution_time
-from hahomematic.exceptions import AuthFailure, BaseHomematicException, NoConnection
-from hahomematic.platforms.device import HmDevice
-from hahomematic.support import (
     ProgramData,
     SystemInformation,
     SystemVariableData,
-    build_headers,
-    build_xml_rpc_uri,
-    get_channel_no,
-    reduce_args,
 )
-from hahomematic.xml_rpc_proxy import XmlRpcProxy
+from hahomematic.exceptions import AuthFailure, BaseHomematicException, NoConnection
+from hahomematic.performance import measure_execution_time
+from hahomematic.platforms.device import HmDevice
+from hahomematic.support import build_headers, build_xml_rpc_uri, get_channel_no, reduce_args
 
-_LOGGER = logging.getLogger("hahomematic.client")
+_LOGGER = logging.getLogger(__name__)
 
 _ADDRESS: Final = "address"
 _CHANNELS: Final = "channels"
@@ -634,7 +629,7 @@ class Client(ABC):
         ):
             _LOGGER.warning(
                 "UPDATE_PARAMSET_DESCRIPTIONS failed: "
-                "Interface missing in central_unit cache. "
+                "Interface missing in central cache. "
                 "Not updating paramsets for %s",
                 device_address,
             )
@@ -644,7 +639,7 @@ class Client(ABC):
         ):
             _LOGGER.warning(
                 "UPDATE_PARAMSET_DESCRIPTIONS failed: "
-                "Channel missing in central_unit.cache. "
+                "Channel missing in central.cache. "
                 "Not updating paramsets for %s",
                 device_address,
             )