+++ hahomematic/client/xml_rpc.py
@@ -11,11 +11,11 @@ from ssl import SSLError
 from typing import Any, Final, TypeVar
 import xmlrpc.client
 
-from hahomematic import central_unit as hmcu
+from hahomematic import central as hmcu
 from hahomematic.exceptions import AuthFailure, ClientException, NoConnection
 from hahomematic.support import get_tls_context, reduce_args
 
-_LOGGER = logging.getLogger("hahomematic.client")
+_LOGGER = logging.getLogger(__name__)
 
 _T = TypeVar("_T")