+++ hahomematic/platforms/custom/__init__.py
@@ -8,7 +8,7 @@ from hahomematic.platforms import device as hmd
 from hahomematic.platforms.custom.definition import entity_definition_exists, get_entity_configs
 from hahomematic.platforms.custom.support import CustomConfig
 
-_LOGGER = logging.getLogger("hahomematic.platform")
+_LOGGER = logging.getLogger(__name__)
 
 
 def create_custom_entity_and_append_to_device(