+++ changelog.md
@@ -1,6 +1,7 @@
 # Version 2023.9.2 (2023-09-15)
 
 - Cleanup logger
+- Refactor client and central modules
 - Move bind_collector to platforms.entity
 - Move config/value_property to platforms.decorators
 - Move measure_execution_time to decorators